examples/with-openai-python/requirements.txt
flexprice>=1.0.17
openai>=1.51.0
python-dotenv>=1.0.1
